anti-debug & anti-tamper

+++ runtime/src/vm/utils/vreg.rs
@@ -5,14 +5,14 @@ use iced_x86::code_asm::{ptr, qword_ptr, AsmRegister32, AsmRegister64};
 pub fn load_reg(rt: &mut Runtime, base: AsmRegister64, src: VMReg, dst: AsmRegister64) {
     // mov ..., [...]
     rt.asm
-        .mov(dst, ptr(base + rt.mapper.index(src) * 8))
+        .mov(dst, ptr(base + rt.mapper.index(src) as i32 * 8))
         .unwrap();
 }
 
 pub fn load_reg32(rt: &mut Runtime, base: AsmRegister64, src: VMReg, dst: AsmRegister32) {
     // mov ..., [...]
     rt.asm
-        .mov(dst, ptr(base + rt.mapper.index(src) * 8))
+        .mov(dst, ptr(base + rt.mapper.index(src) as i32 * 8))
         .unwrap();
 }
 
@@ -25,7 +25,7 @@ pub fn load_mem(
 ) {
     // mov ..., [...]
     rt.asm
-        .mov(tmp, ptr(base + rt.mapper.index(src) * 8))
+        .mov(tmp, ptr(base + rt.mapper.index(src) as i32 * 8))
         .unwrap();
     // mov ..., [...]
     rt.asm.mov(dst, ptr(tmp)).unwrap();
@@ -34,21 +34,21 @@ pub fn load_mem(
 pub fn store_reg(rt: &mut Runtime, base: AsmRegister64, src: AsmRegister64, dst: VMReg) {
     // mov [...], ...
     rt.asm
-        .mov(ptr(base + rt.mapper.index(dst) * 8), src)
+        .mov(ptr(base + rt.mapper.index(dst) as i32 * 8), src)
         .unwrap();
 }
 
 pub fn store_reg32(rt: &mut Runtime, base: AsmRegister64, src: AsmRegister32, dst: VMReg) {
     // mov [...], ...
     rt.asm
-        .mov(ptr(base + rt.mapper.index(dst) * 8), src)
+        .mov(ptr(base + rt.mapper.index(dst) as i32 * 8), src)
         .unwrap();
 }
 
 pub fn store_imm(rt: &mut Runtime, base: AsmRegister64, src: i32, dst: VMReg) {
     // mov [...], ...
     rt.asm
-        .mov(qword_ptr(base + rt.mapper.index(dst) * 8), src)
+        .mov(qword_ptr(base + rt.mapper.index(dst) as i32 * 8), src)
         .unwrap();
 }
 
@@ -61,7 +61,7 @@ pub fn store_mem(
 ) {
     // mov ..., [...]
     rt.asm
-        .mov(tmp, ptr(base + rt.mapper.index(dst) * 8))
+        .mov(tmp, ptr(base + rt.mapper.index(dst) as i32 * 8))
         .unwrap();
     // mov [...], ...
     rt.asm.mov(ptr(tmp), src).unwrap();
@@ -70,69 +70,69 @@ pub fn store_mem(
 pub fn add_reg(rt: &mut Runtime, base: AsmRegister64, src: AsmRegister64, dst: VMReg) {
     // add [...], ...
     rt.asm
-        .add(ptr(base + rt.mapper.index(dst) * 8), src)
+        .add(ptr(base + rt.mapper.index(dst) as i32 * 8), src)
         .unwrap();
 }
 
 pub fn add_imm(rt: &mut Runtime, base: AsmRegister64, src: i32, dst: VMReg) {
     // add [...], ...
     rt.asm
-        .add(qword_ptr(base + rt.mapper.index(dst) * 8), src)
+        .add(qword_ptr(base + rt.mapper.index(dst) as i32 * 8), src)
         .unwrap();
 }
 
 pub fn reg_add(rt: &mut Runtime, base: AsmRegister64, src: VMReg, dst: AsmRegister64) {
     // add ..., [...]
     rt.asm
-        .add(dst, ptr(base + rt.mapper.index(src) * 8))
+        .add(dst, ptr(base + rt.mapper.index(src) as i32 * 8))
         .unwrap();
 }
 
 pub fn sub_reg(rt: &mut Runtime, base: AsmRegister64, src: AsmRegister64, dst: VMReg) {
     // sub [...], ...
     rt.asm
-        .sub(ptr(base + rt.mapper.index(dst) * 8), src)
+        .sub(ptr(base + rt.mapper.index(dst) as i32 * 8), src)
         .unwrap();
 }
 
 pub fn sub_imm(rt: &mut Runtime, base: AsmRegister64, src: i32, dst: VMReg) {
     // sub [...], ...
     rt.asm
-        .sub(qword_ptr(base + rt.mapper.index(dst) * 8), src)
+        .sub(qword_ptr(base + rt.mapper.index(dst) as i32 * 8), src)
         .unwrap();
 }
 
 pub fn reg_sub(rt: &mut Runtime, base: AsmRegister64, src: VMReg, dst: AsmRegister64) {
     // sub ..., [...]
     rt.asm
-        .sub(dst, ptr(base + rt.mapper.index(src) * 8))
+        .sub(dst, ptr(base + rt.mapper.index(src) as i32 * 8))
         .unwrap();
 }
 
 pub fn cmp_reg(rt: &mut Runtime, base: AsmRegister64, dst: VMReg, src: AsmRegister64) {
     // cmp [...], ...
     rt.asm
-        .cmp(ptr(base + rt.mapper.index(dst) * 8), src)
+        .cmp(ptr(base + rt.mapper.index(dst) as i32 * 8), src)
         .unwrap();
 }
 
 pub fn cmp_imm(rt: &mut Runtime, base: AsmRegister64, dst: VMReg, src: i32) {
     // cmp [...], ...
     rt.asm
-        .cmp(qword_ptr(base + rt.mapper.index(dst) * 8), src)
+        .cmp(qword_ptr(base + rt.mapper.index(dst) as i32 * 8), src)
         .unwrap();
 }
 
 pub fn push(rt: &mut Runtime, base: AsmRegister64, src: VMReg) {
     // push [...]
     rt.asm
-        .push(qword_ptr(base + rt.mapper.index(src) * 8))
+        .push(qword_ptr(base + rt.mapper.index(src) as i32 * 8))
         .unwrap();
 }
 
 pub fn pop(rt: &mut Runtime, base: AsmRegister64, dst: VMReg) {
     // pop [...]
     rt.asm
-        .pop(qword_ptr(base + rt.mapper.index(dst) * 8))
+        .pop(qword_ptr(base + rt.mapper.index(dst) as i32 * 8))
         .unwrap();
 }
